Move `get_serializer_for_model` to `utils.api`
