+++ utils/functions.py
@@ -5,8 +5,6 @@ import json
 from django.core.serializers import serialize
 from taggit.managers import _TaggableManager
 
-from peering_manager.api.exceptions import SerializerNotFound
-
 
 # Shamlessly stolen from django.utils.functional (<3.0)
 def curry(_curried_func, *args, **kwargs):
@@ -40,29 +38,6 @@ def generate_signature(data, secret):
     """
     signature = hmac.new(key=secret.encode("utf8"), msg=data, digestmod=hashlib.sha512)
     return signature.hexdigest()
-
-
-def get_serializer_for_model(model, prefix="", suffix=""):
-    """
-    Returns the appropriate API serializer for a model.
-    """
-    app_name, model_name = model._meta.label.split(".")
-    if app_name == "auth":
-        app_name = "users"
-    serializer_name = (
-        f"{app_name}.api.serializers.{prefix}{model_name}{suffix}Serializer"
-    )
-    try:
-        # Try importing the serializer class
-        components = serializer_name.split(".")
-        mod = __import__(components[0])
-        for c in components[1:]:
-            mod = getattr(mod, c)
-        return mod
-    except AttributeError:
-        raise SerializerNotFound(
-            f"Could not determine serializer for {app_name}.{model_name} with prefix '{prefix}' and suffix '{suffix}'"
-        )
 
 
 def is_taggable(instance):
